+++ src/components/containers/BulmaDropdown.vue
@@ -1,19 +1,20 @@
 <template>
+	<component :is="tag" ref="dropdown" :class="[classes]" class="dropdown">
 
-	<component :is="tag" ref="dropdown" :class="[classes,getLeftRightClasses(alignment)]" class="dropdown">
-		<span :is="tag" class="dropdown-trigger" @click="toggleActivation($event, $refs.dropdown, onClick)">
-			<slot/>
+		<span :is="tag" @click="toggleActivation($event, $refs.dropdown,true)">
+			<slot />
 		</span>
 
 		<component :is="tag" class="dropdown-menu">
 			<component :is="tag" class="dropdown-content">
 
 				<component :is="itemTag" v-for="(item,i) in content" ref="item" class="dropdown-item"
-				           @click="emit('itemClick', item,$event, $refs.item)" :key="item+'-'+i">
+				           @click="emit('itemClick', item,$event, $refs.item[i])">
 					<!--!	todo: add better values for a divider-->
-					<hr v-if="item === null || item === undefined " class="dropdown-divider" />
-					<slot v-else-if="$slots['dropdown-item']" name="dropdown-item" v-bind:data="item" />
-					<component :is="tag" v-else class="dropdown-item">{{ item }}</component>
+					<hr v-if="item === null" class="dropdown-divider" />
+					<slot v-else name="dropdown-item" v-bind:data="item">
+						<component :is="tag" class="dropdown-item">{{ item }}</component>
+					</slot>
 				</component>
 
 			</component>
@@ -23,53 +24,34 @@
 
 <script lang="ts" setup>
 
-	import {PropType, ref} from 'vue';
-	import {getBulmaClassesFromProps, getLeftRightClasses, LeftRight, toggleActivation} from '../../types';
+	import {computed, ref} from 'vue';
+	import {BulmaLeftRight, getBulmaClassesFromProps, toggleActivation} from '../../types';
 
 	const emit = defineEmits<{
 		(name: 'itemClick', item: any, event: Event, element: HTMLElement): void
 	}>();
-	//?! wierd stuff happens  in the with defaults macro when wrapping the props in quotes like the 'onClick' not showing up
-	// and 'is-hoverable' coming up as an attribute
-	const props = defineProps({
-		alignment: {
-			type: String as PropType<LeftRight>
-		},
-		content: {
-			/// The content of the dropdown; null/undefined/NaN will be a divider
-			type: Array as PropType<any[]>,
-			default: [],
-			required: true
-		},
-		/// The tag to use for the dropdown,
-		tag: {
-			type: String as PropType<string>,
-			default: 'div'
-		},
-		/// The tag to use for the dropdown items
-		itemTag: {
-			type: String as PropType<string>,
-			default: 'p'
-		},
-		onClick: {
-			type: Boolean,
-			default: true
-		},
-		'is-hoverable': {
-			type: Boolean as PropType<boolean>,
-			default: false
-		},
-		'is-dropup': {
-			type: Boolean as PropType<boolean>,
-			default: false
-		}
+
+	const props = withDefaults(defineProps<{
+		/** alignment of the dropdown menu */
+		alignment?: BulmaLeftRight,
+		/** items to display in the dropdown menu */
+		content: any[],
+		/** tag to use for the dropdown menu */
+		tag?: string,
+		/** tag to use for the dropdown items */
+		itemTag?: string,
+		/** open the dropdown menu on hover */
+		isHoverable?: boolean,
+		/** open the dropdown upwards */
+		isDropup?: boolean
+	}>(), {
+		tag: 'div',
+		itemTag: 'p'
 	});
-
-
-	const classes = getBulmaClassesFromProps(props);
+	const classes = computed(() => getBulmaClassesFromProps(props));
 
 	const dropdown = ref<HTMLElement | null>(null);
-	const item = ref<HTMLElement | null>(null);
+	const item = ref<HTMLElement[] | null>(null);
 
 </script>